fixes - editor fast render + folder renames

+++ src/global_state/mod.rs
@@ -309,7 +309,7 @@ impl GlobalState {
     pub fn full_resize(&mut self, height: u16, width: u16, workspace: &mut Workspace) {
         self.screen_rect = (width, height).into();
         self.draw_callback = draw::full_rebuild;
-        workspace.resize_all(self.editor_area.width as usize, self.editor_area.height as usize);
+        workspace.resize_all(self.editor_area.width, self.editor_area.height as usize);
     }
 
     pub fn recalc_draw_size(&mut self) {
